pyproject.toml
[project]
name = "mcp-client-langchain-py"
version = "0.0.1"
description = "MCP Client Using LangChain / Python"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "jsonschema-pydantic>=0.6",
    "langchain>=0.3.14",
    "langchain-anthropic>=0.3.1",
    "langchain-groq>=0.2.3",
    "langchain-openai>=0.3.0",
    "langgraph>=0.2.62",
    "mcp>=1.2.0",
    "pyjson5>=1.6.8",
    "pympler>=1.1",
    "python-dotenv>=1.0.1",
]
